Data for Genotypes by Environment Interaction (GEI)

cultivo2008 is used for performing Genotypes by Environment Interaction (GEI) Analysis.

Usage

data(cultivo2008)

Format

A data.frame 1320 obs. of 6 variables.

Details

Gen Genotype

Institute Institute

Rep Replicate

Block Block

Env Environment

Yield Yield Response

References

Perez-Elizalde, S., Jarquin, D., and Crossa, J. (2011) A General Bayesian Estimation Method of Linear–Bilinear Models Applied to Plant Breeding Trials With Genotype × Environment Interaction. Journal of Agricultural, Biological, and Environmental Statistics, 17, 15–37. (doi:10.1007/s13253-011-0063-9)

Author

Muhammad Yaseen (myaseen208@gmail.com)

Diego Jarquin (diego.jarquin@gmail.com)

Sergio Perez-Elizalde (sergiop@colpos.mx)

Juan Burgueño (j.burgueno@cgiar.org)

Jose Crossa (j.crossa@cgiar.org)

Examples

data(cultivo2008)